refactor: useCampaignGroup.ts 생성 및 분리
